Raise player limit from 8 to 32 and wire up start position service
Decouples civilization assignment from player index beyond the original
7-slot "preferred player number" pairing (CivilizationAssignment), widens
the player colour palette and the tile Visited bitmask, fixes two latent
bugs where respawn/replay data used Civilization.PreferredPlayerNumber
instead of the actual player slot, and raises the NewGame competition
selection up to 31 non-barbarian players. The legacy SVE save format
remains capped at 8 players via the existing SveSaveCompatibilityService
gate; YAML saves support the full range.

Also includes the new start position placement services
(AreaBasedStartPositionService/LegacyStartPositionService) that
Game.NewGame.cs now depends on for placing starting units.

+++ src/Tiles/ITile.cs
@@ -54,7 +54,7 @@ namespace CivOne.Tiles
 		byte FortressCost { get; set; }
 		bool Mine { get; set; }
 		bool Hut { get; set; }
-		byte Visited { get; }
+		uint Visited { get; }
 		void Visit(byte owner);
 		bool IsOcean { get; }
 		City City { get; }
